Move classes into duke package
