+++ src/main/java/duke/Deadline.java
@@ -1,46 +1,48 @@
+package duke;
+
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
 
-public class Event extends Task {
-    protected LocalDate at;
+public class Deadline extends Task {
+    protected LocalDate by;
     /**
-     * Constructor for a event instance.
+     * Constructor for a deadline instance.
      *
-     * @param description the description of the event
-     * @param at the time of the event
+     * @param description the description of the task
+     * @param by the due date of the task
      */
-    public Event(String description, String at) throws DukeException {
+    public Deadline(String description, String by) throws DukeException {
         super(description);
         try {
-            this.at = LocalDate.parse(at);
+            this.by = LocalDate.parse(by);
         } catch (java.time.format.DateTimeParseException e) {
             throw new DukeException("Invalid Date Format, please input it as YYYY-MM-DD");
         }
     }
 
     /**
-     * String representation of the event.
+     * String representation of the deadline.
      *
-     * @return String representing this event
+     * @return String representing this task
      */
     @Override
     public String toString() {
-        return "[E]" +
+        return "[D]" +
                 super.toString() +
-                " (at: " +
-                this.at.format(DateTimeFormatter.ofPattern("MMM d yyyy")) +
+                " (by: " +
+                this.by.format(DateTimeFormatter.ofPattern("MMM d yyyy")) +
                 ")";
     }
 
     /**
-     * Gets the deadline of the Task, if any.
+     * Gets the deadline of the duke.Task, if any.
      *
-     * @return A LocalDate representing the Task's deadline if it exists
+     * @return A LocalDate representing the duke.Task's deadline if it exists
      *         null if no such deadline exists
      */
     @Override
     public LocalDate getDeadline() {
-        return this.at;
+        return this.by;
     }
 
     /**
@@ -51,8 +53,8 @@ public class Event extends Task {
      */
     public String encode() {
         return String.format("%s # %s # %s",
-                "E",
+                "D",
                 super.encode(),
-                this.at);
+                this.by);
     }
 }
